week_3/day_3/chatbot-educatif/requirements.txt
openai
pymongo
python-dotenv
requests
typing-extensions
rich  # optionnel
